@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.25.0-devel
-// 	protoc        v3.17.3
+// 	protoc-gen-go v1.28.1
+// 	protoc        v3.21.9
 // source: internal/plugin/proto/rule.proto
 
 package proto
