@@ -1,4 +1,8 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
+// versions:
+// - protoc-gen-go-grpc v1.2.0
+// - protoc             v3.21.9
+// source: internal/plugin/proto/rule.proto
 
 package proto
 
@@ -11,6 +15,7 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
+// Requires gRPC-Go v1.32.0 or later.
 const _ = grpc.SupportPackageIsVersion7
 
 // HCLvetRulePluginClient is the client API for HCLvetRulePlugin service.
@@ -76,7 +81,7 @@ type UnsafeHCLvetRulePluginServer interface {
 }
 
 func RegisterHCLvetRulePluginServer(s grpc.ServiceRegistrar, srv HCLvetRulePluginServer) {
-	s.RegisterService(&_HCLvetRulePlugin_serviceDesc, srv)
+	s.RegisterService(&HCLvetRulePlugin_ServiceDesc, srv)
 }
 
 func _HCLvetRulePlugin_GetRuleInfo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
@@ -115,7 +120,10 @@ func _HCLvetRulePlugin_ExecuteRule_Handler(srv interface{}, ctx context.Context,
 	return interceptor(ctx, in, info, handler)
 }
 
-var _HCLvetRulePlugin_serviceDesc = grpc.ServiceDesc{
+// HCLvetRulePlugin_ServiceDesc is the grpc.ServiceDesc for HCLvetRulePlugin service.
+// It's only intended for direct use with grpc.RegisterService,
+// and not to be introspected or modified (even as a copy)
+var HCLvetRulePlugin_ServiceDesc = grpc.ServiceDesc{
 	ServiceName: "proto.HCLvetRulePlugin",
 	HandlerType: (*HCLvetRulePluginServer)(nil),
 	Methods: []grpc.MethodDesc{
